우승자 결정로직이 이동하면서 테스트가 더 명확해지고 쉬워짐
